+++ src/test/java/io/lettuce/core/commands/reactive/HashReactiveCommandIntegrationTests.java
@@ -15,8 +15,18 @@
  */
 package io.lettuce.core.commands.reactive;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
+import java.util.stream.Collectors;
+
 import javax.inject.Inject;
 
+import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.api.Test;
+
+import reactor.test.StepVerifier;
+import io.lettuce.core.KeyValue;
+import io.lettuce.core.Value;
 import io.lettuce.core.api.StatefulRedisConnection;
 import io.lettuce.core.commands.HashCommandIntegrationTests;
 import io.lettuce.test.ReactiveSyncInvocationHandler;
@@ -26,8 +36,31 @@ import io.lettuce.test.ReactiveSyncInvocationHandler;
  */
 class HashReactiveCommandIntegrationTests extends HashCommandIntegrationTests {
 
+    private final StatefulRedisConnection<String, String> connection;
+
     @Inject
     HashReactiveCommandIntegrationTests(StatefulRedisConnection<String, String> connection) {
         super(ReactiveSyncInvocationHandler.sync(connection));
+        this.connection = connection;
+    }
+
+    @Test
+    public void hgetall() {
+
+        connection.sync().hset(key, "zero", "0");
+        connection.sync().hset(key, "one", "1");
+        connection.sync().hset(key, "two", "2");
+
+        connection.reactive().hgetall(key).collect(Collectors.toMap(KeyValue::getKey, Value::getValue)).as(StepVerifier::create)
+                .assertNext(actual -> {
+
+                    assertThat(actual).containsEntry("zero", "0").containsEntry("one", "1").containsEntry("two", "2");
+                }).verifyComplete();
+    }
+
+    @Test
+    @Disabled("API differences")
+    public void hgetallStreaming() {
+
     }
 }